+++ client/actions/actionTypes.js
@@ -1,0 +1,4 @@
+export const GET_PRODUCTS = 'GET_PRODUCTS';
+export const GET_SINGLE_PRODUCT = 'GET_SINGLE_PRODUCT';
+export const GET_CATEGORIES = 'GET_CATEGORIES';
+export const GET_SINGLE_CATEGORY = 'GET_SINGLE_CATEGORY';
