+++ client/index.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Main from './components/Main.js';
-import store from './components/redux/store.js';
+import Main from './components/Main';
+import store from './store';
 
 const appDiv = document.getElementById('app');
